requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
Pillow
pymupdf
pandas
openpyxl
